fix: rename file, improve logic
