+++ src/index.js
@@ -1,3 +1,3 @@
 import "./style.css";
-// import * as gameLogic from "./game-logic";
-const {computer} = require('./computer')
+import computer from "./dom-ui";
+// import human from "./human";
